Add FastAPI server and refactor automation runner
