+++ ai_testing_tool/__init__.py
@@ -1,0 +1,5 @@
+"""Core package for the AI testing tool."""
+
+from .runner import RunResult, run_tasks, main
+
+__all__ = ["RunResult", "run_tasks", "main"]
